+++ app/Http/Controllers/FileController.php
@@ -3,6 +3,7 @@
 namespace Servidor\Http\Controllers;
 
 use Illuminate\Http\Request;
+use InvalidArgumentException;
 use Servidor\FileManager\FileManager;
 
 class FileController extends Controller
@@ -27,11 +28,27 @@ class FileController extends Controller
     public function index(Request $request)
     {
         if ($filepath = $request->get('file')) {
-            return $this->fm->open($filepath);
+            $file = $this->fm->open($filepath);
+
+            if (array_key_exists('error', $file)) {
+                return response($file, $file['error']['code']);
+            }
+
+            try {
+                return response()->json($file);
+            } catch (InvalidArgumentException $e) {
+                $file['contents'] = '';
+                $file['error'] = [
+                    'code' => 422,
+                    'msg' => 'Failed loading file',
+                ];
+
+                return response()->json($file);
+            }
         }
 
         $path = $request->get('path');
 
-        return $this->fm->list($path);
+        return response()->json($this->fm->list($path));
     }
 }