+++ app/StatsBar.php
@@ -10,6 +10,8 @@ class StatsBar
         $lsb = self::parseReleaseFile('lsb');
 
         return [
+            'cpu' => self::getCpuUsage(),
+            'ram' => self::getRamUsage(),
             'hostname' => gethostname(),
             'os' => [
                 'name' => php_uname('s'),
@@ -19,7 +21,7 @@ class StatsBar
         ];
     }
 
-    protected static function parseReleaseFile($file)
+    private static function parseReleaseFile($file)
     {
         $flags = FILE_IGNORE_NEW_LINES;
         $data = [];
@@ -34,5 +36,29 @@ class StatsBar
         }
 
         return $data;
+    }
+
+    /**
+     * Get the current CPU usage in percent.
+     */
+    private static function getCpuUsage(): float
+    {
+        return (float) exec("mpstat | tail -n1 | awk '{ print 100 - $12 }'");
+    }
+
+    /**
+     * Get details about the RAM currently used/free.
+     */
+    private static function getRamUsage(): array
+    {
+        $output = exec('free | tail -n+2 | head -n1');
+
+        $data = sscanf($output, '%s %d %d %d %d %d');
+
+        return [
+            'total' => round($data[1] / 1024),
+            'used' => round($data[2] / 1024),
+            'free' => round(($data[3] + $data[5]) / 1024),
+        ];
     }
 }