+++ resources/js/components/Sites/SiteItem.vue
@@ -4,10 +4,16 @@
             <sui-card-header>{{ site.name }}</sui-card-header>
             <sui-card-meta>{{ site.primary_domain }}</sui-card-meta>
         </sui-card-content>
-        <router-link is="sui-button" attached="bottom"
-            :to="{ name: 'apps.edit', params: { id: site.id } }">
-            <sui-icon name="cogs"></sui-icon> Manage Site
-        </router-link>
+
+        <sui-card-content extra>
+            <router-link :to="{ name: 'apps.edit', params: { id: site.id } }">
+                <sui-icon name="cogs"></sui-icon> Manage Site
+            </router-link>
+            <router-link :to="{ name: 'files', params: { path: site.document_root } }"
+                slot="right" v-if="site.document_root != null">
+                <sui-icon name="open folder" /> Browse Files
+            </router-link>
+        </sui-card-content>
     </sui-card>
 </template>