+++ CHANGELOG.md
@@ -9,6 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.4.0] - 2019-08-24
+### Added
+* Buttons to navigate between Sites and their files
+* CPU usage and current free RAM added to Stats Bar
+
+### Changed
+* Icons in the File Manager are now coloured based on type (file or folder)
+* Error handling when opening files in File Manager is greatly improved
+* Segments in the File Browser's path bar will now jump to that path
+
+### Fixed
+* Path wouldn't update in File Manager URL while changing folders
+* Domain name is now required when updating a Site
+
+
 ## [0.3.0] - 2019-08-13
 ### Added
 * File Manager with basic text file viewing capability
@@ -32,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Basic login/logout auth stuff
 
 
-[Unreleased]: https://github.com/dshoreman/servidor/compare/v0.3.0...develop
+[Unreleased]: https://github.com/dshoreman/servidor/compare/v0.4.0...develop
+[0.4.0]: https://github.com/dshoreman/servidor/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/dshoreman/servidor/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/dshoreman/servidor/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/dshoreman/servidor/releases/tag/v0.1.0